Diagrama, rename de packages
