+++ app/src/main/kotlin/ayds/lisboa/songinfo/moredetails/presentation/OtherInfoUiState.kt
@@ -1,4 +1,4 @@
-package ayds.lisboa.songinfo.moredetails.presenter
+package ayds.lisboa.songinfo.moredetails.presentation
 
 const val LAST_FM_DEFAULT_IMAGE = "https://upload.wikimedia.org/wikipedia/commons/thumb/d/d4/Lastfm_logo.svg/320px-Lastfm_logo.svg.png"
